Refactor : ErrorCode를 exception 패키지 내부로 이동
