+++ ddv/src/main/java/community/ddv/domain/board/service/DiscussionService.java
@@ -1,6 +1,6 @@
 package community.ddv.domain.board.service;
 
-import community.ddv.global.constant.ErrorCode;
+import community.ddv.global.exception.ErrorCode;
 import community.ddv.domain.board.dto.ReviewDTO;
 import community.ddv.domain.board.dto.ReviewDTO.ReviewUpdateDTO;
 import community.ddv.domain.board.dto.ReviewResponseDTO;
