+++ ddv/src/main/java/community/ddv/domain/vote/service/VoteService.java
@@ -1,7 +1,7 @@
 package community.ddv.domain.vote.service;
 
 import community.ddv.domain.movie.service.MovieService;
-import community.ddv.global.constant.ErrorCode;
+import community.ddv.global.exception.ErrorCode;
 import community.ddv.global.constant.Role;
 import community.ddv.domain.movie.dto.MovieDTO;
 import community.ddv.domain.vote.dto.VoteDTO.VoteCreatedDTO;
